plugin_based_callback/requirements.txt
cryptography>=42.0.2
motor>=3.3.2
pymongo>=4.6.1
asyncpg>=0.29.0
fastapi>=0.109.2
python-dotenv>=1.0.1
PyJWT>=2.8.0
uvicorn>=0.27.0.post1
aiofiles>=23.2.1